Big Rewrite

+++ uninstall.php
@@ -1,5 +1,5 @@
 <?php
-defined('WP_UNINSTALL_PLUGIN') || die();
-
-delete_option('ww_yubico_otp');
-delete_metadata('user', 0, '_ww_yotp', null, true);
+if ( defined( 'WP_UNINSTALL_PLUGIN' ) ) {
+	delete_option( 'ww_yubico_otp' );
+	delete_metadata( 'user', 0, '_ww_yotp', null, true );
+}
